refactor extract to status reporter class
